fix: int 범위를 벗어난 다리 길이 검증을 위해 Long타입으로 변환하여 에러처리함

+++ src/main/java/bridge/domain/BridgeSize.java
@@ -13,14 +13,14 @@ public class BridgeSize {
 
     private void validBridgeSize(String bridgeSize) {
         validIsDigit(bridgeSize);
-        validBridgeSizeRange(Integer.parseInt(bridgeSize));
+        validBridgeSizeRange(Long.parseLong(bridgeSize));
     }
     private void validIsDigit(String bridgeSize) {
         if (!bridgeSize.chars().allMatch(Character::isDigit)) {
             throw new IllegalArgumentException(ErrorMessage.NOT_NUMBER.getMessage());
         }
     }
-    private void validBridgeSizeRange(int bridgeSize) {
+    private void validBridgeSizeRange(long bridgeSize) {
         if (bridgeSize< Constant.BRIDGE_SIZE_MIN || bridgeSize>Constant.BRIDGE_SIZE_MAX) {
             throw new IllegalArgumentException(ErrorMessage.NOT_BRIDGE_SIZE_RANGE.getMessage());
         }
